// src/api/v1/types/cart/cartDTO.ts
// src/types/cart/cartDTO.ts
import mongoose from 'mongoose';

export interface CartItemInput {
  inventoryId: mongoose.Types.ObjectId | string;
  productId: mongoose.Types.ObjectId | string;
  colorId?: mongoose.Types.ObjectId | string;
  size?: string;
  quantity: number;
}

export interface AddToCartInput extends CartItemInput {}
export interface UpdateCartItemInput {
  inventoryId: mongoose.Types.ObjectId | string;
  quantity: number;
}

export interface RemoveFromCartInput {
  inventoryId: mongoose.Types.ObjectId | string;
}